fix: Add streaming support for Kimi and DeepSeek reasoning models
## Problem
- Kimi 2T (thinking model) was bypassing streaming infrastructure, causing:
  - Empty progress view during reasoning
  - Incomplete final responses (missing content/tool calls)
  - Tool calls not being aggregated properly
- DeepSeek had duplicate aggregator code (~190 lines)
- Legacy function_call support still present

## Solution
1. Created BaseReasoningStreamAggregator
   - Shared base class for all OpenAI-compatible reasoning models
   - Properly aggregates content, reasoning, and tool calls during streaming
   - Removed deprecated function_call support
   - Handles parallel tool calls via index field

2. Updated Kimi Handler
   - Added streaming support for thinking models with aggregator
   - Fixed tool calls by adding tools parameter to kwargs
   - Properly extracts reasoning_content from stream chunks
   - Supports both streaming and non-streaming modes

3. Updated DeepSeek Handler
   - Now uses BaseReasoningStreamAggregator
   - Removed ~190 lines of duplicate code
   - Removed legacy function_call support

4. Added Package Configs
   - texra.model.useStreamingMoonshot (default: true)
   - texra.model.useStreamingDashscope (default: true)

## Benefits
- Single source of truth for reasoning model streaming
- Reduced code duplication (~185 lines eliminated)
- Tool calls now work correctly for Kimi models
- Future reasoning models can reuse the base class
- Cleaner, more maintainable codebase
